requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
google-cloud-speech==2.26.0
python-multipart==0.0.9
protobuf